requirements.txt
# Core API clients
openai>=1.0.0
anthropic>=0.18.0
google-generativeai>=0.3.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# Statistical analysis
scipy>=1.10.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0

# Optional: for advanced analysis
# statsmodels>=0.14.0
# pingouin>=0.5.3